Trait julia_set::Backend

pub trait Backend<In>: Default {
    type Error: Debug + Display;
    type Program: Render;
    fn create_program(&self, function: In) -> Result<Self::Program, Self::Error>;
}

Backend capable of converting an input (the type parameter) into a program. The program then can be used to Render the Julia set with various rendering Params.

Associated Types

type Error: Debug + Display

Error that may be returned during program creation.

type Program: Render

Program output by the backend.

Loading content...
Required methods

fn create_program(&self, function: In) -> Result<Self::Program, Self::Error>

Creates a program.

Errors

May return an error if program cannot be created (out of memory, etc.).

Loading content...