Trait julia_set::Backend

pub trait Backend<In>: Default {
    type Error: Debug + Display;
    type Program: Render;

    // Required method
    fn create_program(&self, function: In) -> Result<Self::Program, Self::Error>;
}

Backend capable of converting an input (the type parameter) into a program. The program then can be used to Render the Julia set with various rendering Params.

Required Associated Types

type Error: Debug + Display

Error that may be returned during program creation.

type Program: Render

Program output by the backend.

Required Methods

fn create_program(&self, function: In) -> Result<Self::Program, Self::Error>

Creates a program.

Errors

May return an error if program cannot be created (out of memory, etc.).

Object Safety

This trait is not object safe.