Module julia_set::transform

Post-processing transforms on Julia set images.

These transforms are not meant to be exhaustive; rather, they represent some functionality lacking from image and imageproc crates. You can absolutely use the transforms from these crates together with (or instead of) transforms defined here.

Examples

use julia_set::transform::{ApplyTransform, Negative, Smoothstep};

let image: GrayImage = // ...
let transformed_image = image.apply(Negative).apply(Smoothstep).transform();

Structs

ImageAndTransform
Source image together with one or more PixelTransforms. Transforms are applied lazily, once Self::transform() is called.

Negative
Transform that negates the value of each pixel of the input image.

Palette
Transform that colorizes the image using a palette.

Smoothstep
Transform that smooths the value of each pixel of the input image using cubic Hermite interpolation (aka smoothstep from GLSL / OpenCL).

Traits

ApplyTransform
Trait allowing to lazily apply one or more PixelTransforms to an image.

PixelTransform
Pixel-wise transform.