Trait julia_set::transform::ApplyTransform

pub trait ApplyTransform<Pix: Pixel, F> {
    type CombinedTransform: PixelTransform<Pix>;
    fn apply(
        self, 
        transform: F
    ) -> ImageAndTransform<Pix, Self::CombinedTransform>;
}

Trait allowing to lazily apply one or more PixelTransforms to an image.

This trait is implemented for [ImageBuffer]s and for ImageAndTransforms, allowing to chain transforms.

Associated Types

type CombinedTransform: PixelTransform<Pix>

Combined transform after applying transform.

Loading content...
Required methods

fn apply(self, transform: F) -> ImageAndTransform<Pix, Self::CombinedTransform>

Appends transform to the list of transforms to be performed on the source image.

Loading content...