Trait julia_set::transform::PixelTransform

pub trait PixelTransform<Pix: Pixel> {
    type Output: Pixel + 'static;

    // Required method
    fn transform_pixel(&self, pixel: Pix) -> Self::Output;
}

Pixel-wise transform.

The function defined by the transform is applied separately to each pixel of the source image. The transform may change the type of pixels (e.g., transform from grayscale to a color image), or may leave it intact.

Required Associated Types

type Output: Pixel + 'static

Pixel type for the output image.

Required Methods

fn transform_pixel(&self, pixel: Pix) -> Self::Output

Performs transform on a single pixel.